refactor: enhance command palette modularity
