+++ web/ce/components/command-palette/modals/index.ts
@@ -1,0 +1,3 @@
+export * from "./workspace-level";
+export * from "./project-level";
+export * from "./issue-level";
