+++ web/core/store/base-command-palette.store.ts
@@ -9,9 +9,8 @@ export interface ModalData {
   viewId: string;
 }
 
-export interface ICommandPaletteStore {
+export interface IBaseCommandPaletteStore {
   // observables
-
   isCommandPaletteOpen: boolean;
   isShortcutModalOpen: boolean;
   isCreateProjectModalOpen: boolean;
@@ -22,8 +21,9 @@ export interface ICommandPaletteStore {
   isCreateIssueModalOpen: boolean;
   isDeleteIssueModalOpen: boolean;
   isBulkDeleteIssueModalOpen: boolean;
+  createIssueStoreType: TCreateModalStoreTypes;
   // computed
-  isAnyModalOpen: boolean;
+  isAnyBaseModalOpen: boolean;
   // toggle actions
   toggleCommandPaletteModal: (value?: boolean) => void;
   toggleShortcutModal: (value?: boolean) => void;
@@ -35,11 +35,9 @@ export interface ICommandPaletteStore {
   toggleCreateModuleModal: (value?: boolean) => void;
   toggleDeleteIssueModal: (value?: boolean) => void;
   toggleBulkDeleteIssueModal: (value?: boolean) => void;
-
-  createIssueStoreType: TCreateModalStoreTypes;
 }
 
-export class CommandPaletteStore implements ICommandPaletteStore {
+export abstract class BaseCommandPaletteStore implements IBaseCommandPaletteStore {
   // observables
   isCommandPaletteOpen: boolean = false;
   isShortcutModalOpen: boolean = false;
@@ -51,7 +49,6 @@ export class CommandPaletteStore implements ICommandPaletteStore {
   isDeleteIssueModalOpen: boolean = false;
   isBulkDeleteIssueModalOpen: boolean = false;
   createPageModal: TCreatePageModal = DEFAULT_CREATE_PAGE_MODAL_DATA;
-
   createIssueStoreType: TCreateModalStoreTypes = EIssuesStoreType.PROJECT;
 
   constructor() {
@@ -67,8 +64,9 @@ export class CommandPaletteStore implements ICommandPaletteStore {
       isDeleteIssueModalOpen: observable.ref,
       isBulkDeleteIssueModalOpen: observable.ref,
       createPageModal: observable,
+      createIssueStoreType: observable,
       // computed
-      isAnyModalOpen: computed,
+      isAnyBaseModalOpen: computed,
       // projectPages: computed,
       // toggle actions
       toggleCommandPaletteModal: action,
@@ -85,20 +83,20 @@ export class CommandPaletteStore implements ICommandPaletteStore {
   }
 
   /**
-   * Checks whether any modal is open or not.
+   * Checks whether any modal is open or not in the base command palette.
    * @returns boolean
    */
-  get isAnyModalOpen() {
+  get isAnyBaseModalOpen() {
     return Boolean(
       this.isCreateIssueModalOpen ||
-        this.isCreateCycleModalOpen ||
-        this.isCreateProjectModalOpen ||
-        this.isCreateModuleModalOpen ||
-        this.isCreateViewModalOpen ||
-        this.isShortcutModalOpen ||
-        this.isBulkDeleteIssueModalOpen ||
-        this.isDeleteIssueModalOpen ||
-        this.createPageModal.isOpen
+      this.isCreateCycleModalOpen ||
+      this.isCreateProjectModalOpen ||
+      this.isCreateModuleModalOpen ||
+      this.isCreateViewModalOpen ||
+      this.isShortcutModalOpen ||
+      this.isBulkDeleteIssueModalOpen ||
+      this.isDeleteIssueModalOpen ||
+      this.createPageModal.isOpen
     );
   }
