+++ web/ee/components/command-palette/modals/index.ts
@@ -1,0 +1,1 @@
+export * from "ce/components/command-palette/modals";
