+++ web/ee/store/command-palette.store.ts
@@ -1,0 +1,1 @@
+export * from "ce/store/command-palette.store";
