+++ web/helpers/command-palette.ts
@@ -1,0 +1,1 @@
+export * from "ce/helpers/command-palette";
